finish migrate text-kit
