+++ packages/text-kit/package.json
@@ -1,34 +1,31 @@
 {
-  "name": "text-kit",
-  "version": "4.3.0",
-  "description": "bundle of useful tools for manipulating text",
-  "author": "Xiaoxing Hu <hi@xiaoxing.dev>",
-  "homepage": "https://github.com/orgapp/orgajs/tree/main/packages/text-kit#readme",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/orgapp/orgajs.git",
-    "directory": "packages/text-kit"
-  },
-  "license": "MIT",
-  "main": "dist/index.js",
-  "type": "module",
-  "directories": {
-    "dist": "dist"
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "clean": "tsc --build --clean",
-    "build": "tsc --build",
-    "test": "node --test --no-warnings --import tsx src/*.test.ts"
-  },
-  "bugs": {
-    "url": "https://github.com/orgapp/orgajs/issues"
-  },
-  "devDependencies": {
-    "@types/unist": "3.0.3",
-    "tsx": "^4.19.2",
-    "typescript": "^5.7.3"
-  }
+	"name": "text-kit",
+	"version": "4.3.0",
+	"description": "bundle of useful tools for manipulating text",
+	"author": "Xiaoxing Hu <hi@xiaoxing.dev>",
+	"homepage": "https://github.com/orgapp/orgajs/tree/main/packages/text-kit#readme",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/orgapp/orgajs.git",
+		"directory": "packages/text-kit"
+	},
+	"license": "MIT",
+	"type": "module",
+	"files": [
+		"lib",
+		"index.js",
+		"index.d.ts",
+		"index.d.ts.map"
+	],
+	"scripts": {
+		"test": "node --test --no-warnings tests/*.js"
+	},
+	"bugs": {
+		"url": "https://github.com/orgapp/orgajs/issues"
+	},
+	"devDependencies": {
+		"@types/unist": "3.0.3",
+		"tsx": "^4.19.2",
+		"typescript": "^5.7.3"
+	}
 }
