--- packages/text-kit/src/utils/index.ts
@@ -1,7 +1,0 @@
-import { CoreAPI } from '../core.js'
-import substring from './substring.js'
-import lines from './lines.js'
-
-export default (core: CoreAPI) => {
-  return substring(lines(core))
-}
