+++ packages/text-kit/tsconfig.json
@@ -1,10 +1,3 @@
 {
-  "extends": "../../tsconfig.esm.json",
-  "compilerOptions": {
-    "noImplicitAny": true,
-    "strictNullChecks": true,
-    "rootDir": "./src",
-    "outDir": "./dist"
-  },
-  "include": ["./src"]
+	"extends": "../../tsconfig.json"
 }
