+++ tsconfig.json
@@ -16,6 +16,7 @@
 	"exclude": ["**/coverage/", "**/node_modules/", "**/tests/"],
 	"include": [
 		"packages/metadata/**/*.js",
+		"packages/text-kit/**/*.js",
 		"packages/reorg-rehype/**/*.js",
 		"packages/orgx/**/*.js",
 		"packages/oast-to-prose/**/*.js",
